refactor(dashboard): extract Court agents lifecycle + credentials (F-B-202 PR-7/10)
Audit ref: imp/audits/2026-05-20/findings/track-b/F-B-202.md

Sprint 2 modularization PR-7 of 10. base=main. **Biggest single
extract of the sprint** — 12 routes, ~917 LOC split into two
sibling sub-routers covering the entire agent management surface.

What moves:

  app/dashboard/agents_lifecycle_routes.py (new, 400 LOC, 5 routes):
    GET  /dashboard/agents/register          register form
    POST /dashboard/agents/register          register submit (sealed-gate)
    GET  /dashboard/agents/{id}/manage       unified settings form
    POST /dashboard/agents/{id}/manage       update profile OR upload-cert
                                             (sealed-gate)
    POST /dashboard/agents/{id}/delete       delete + revoke binding
                                             (sealed-gate)

  app/dashboard/agents_credentials_routes.py (new, 517 LOC, 7 routes):
    GET  /dashboard/agents/{id}              developer portal detail page
    GET  /dashboard/agents/{id}/upload-cert  cert upload form
    POST /dashboard/agents/{id}/upload-cert  cert upload submit (sealed-gate)
    POST /dashboard/agents/{id}/credentials  credentials zip download (sealed-gate)
    POST /dashboard/agents/{id}/bundle       full deploy bundle download (sealed-gate)
    GET  /dashboard/agents/{id}/rotate-cert  cert rotate form
    POST /dashboard/agents/{id}/rotate-cert  cert rotate submit (sealed-gate)

  app/dashboard/router.py (2372 → 1611 LOC, -761):
    - 12 routes removed; surgical cut via py-script (lines 1299-2143
      replaced with breadcrumb comment).
    - Two new include_router calls in the wiring block.

Behavioural delta: zero. All sealed-gate semantics
(_require_sealed_reauth + _sealed_mutation_details audit log)
preserved.

Constants duplicated into ``agents_lifecycle_routes.py`` for self-
containment: _CAPABILITY_MAX_LEN = 64, _CAPABILITY_MAX_COUNT = 50.
Future micro-refactor can promote them via _helpers.py.

Tests:
  - pytest tests/test_dashboard*.py -q — 202/203 (1 pre-existing deselect)
  - pytest tests/ <full suite> -q — 4216 passed, 9 skipped, 32 xfailed
    (1 xdist flake pre-existing, passes isolated)

Conflict warning: base=main (post-PR-4 only on origin). When PR-5 #848
+ PR-6 #849 merge, this PR rebases clean (different sections — agents
sections 22+23+24 vs orgs sections 9+10+19+21+18+20).

Router LOC progress:
  Pre-Sprint 2 : 67 routes, 3125 LOC
  Post-PR-1 #839: 67 routes, 2930 LOC
  Post-PR-2 #841: 60 routes, 2614 LOC
  Post-PR-3 #842: 57 routes, 2528 LOC
  Post-PR-4 #847: 50 routes, 2372 LOC
  Post-PR-5 #848 (still open): 39 routes, 2071 LOC
  Post-PR-6 #849 (still open): 32 routes, ~1828 LOC
  Post-PR-7 (this, before #848+#849 rebase): 38 routes, 1611 LOC
  Post-PR-5+6+7 merged: 20 routes, ~1310 LOC
  Target post-PR-10: ~3 (aggregator), ~100 LOC

Past 70% on both metrics — 47/67 routes extracted, 58% LOC trimmed.

Next PR-8: extract policies_routes.py + bindings_routes.py
(sections 25+26, ~302 LOC, 8 routes).
