+++ app/dashboard/router.py
@@ -112,6 +112,14 @@ from app.dashboard import org_seal_routes as _org_seal_routes  # noqa: E402
 router.include_router(_org_onboard_routes.router)
 router.include_router(_org_seal_routes.router)
 
+# F-B-202 PR-7: include agents_lifecycle (register / manage / delete)
+# and agents_credentials (detail / upload-cert / credentials / bundle
+# / rotate-cert). 12 routes total — biggest extract of Sprint 2.
+from app.dashboard import agents_lifecycle_routes as _agents_lifecycle_routes  # noqa: E402
+from app.dashboard import agents_credentials_routes as _agents_credentials_routes  # noqa: E402
+router.include_router(_agents_lifecycle_routes.router)
+router.include_router(_agents_credentials_routes.router)
+
 
 # ─────────────────────────────────────────────────────────────────────────────
 # Login / Logout
@@ -606,850 +614,12 @@ async def rfq_approve(request: Request, rfq_id: str, db: AsyncSession = Depends(
 # F-B-202 PR-5.
 
 
-# ─────────────────────────────────────────────────────────────────────────────
-# Register Agent
-# ─────────────────────────────────────────────────────────────────────────────
+# Agent register / manage / delete routes moved to
+# ``app/dashboard/agents_lifecycle_routes.py`` since F-B-202 PR-7.
+# Agent detail / upload-cert / credentials / bundle / rotate-cert
+# routes moved to ``app/dashboard/agents_credentials_routes.py`` since
+# F-B-202 PR-7.
 
-@router.get("/agents/register", response_class=HTMLResponse)
-async def agent_register_form(request: Request, db: AsyncSession = Depends(get_db)):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-
-    q = select(OrganizationRecord).where(OrganizationRecord.status == "active").order_by(OrganizationRecord.org_id)
-    orgs = (await db.execute(q)).scalars().all()
-
-    return templates.TemplateResponse("agent_register.html",
-        _ctx(request, session, active="agents", form={}, orgs=orgs, error=None, success=None))
-
-
-@router.post("/agents/register", response_class=HTMLResponse)
-async def agent_register_submit(request: Request, db: AsyncSession = Depends(get_db)):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    if not await verify_csrf(request, session):
-        q = select(OrganizationRecord).where(OrganizationRecord.status == "active").order_by(OrganizationRecord.org_id)
-        orgs = (await db.execute(q)).scalars().all()
-        return templates.TemplateResponse("agent_register.html",
-            _ctx(request, session, active="agents", form={}, orgs=orgs,
-                 error="Invalid CSRF token. Please try again.", success=None),
-            status_code=403)
-    form_data = await request.form()
-    org_id      = form_data.get("org_id", "").strip()
-    agent_name  = form_data.get("agent_name", "").strip()
-    display_name = form_data.get("display_name", "").strip()
-    description  = form_data.get("description", "").strip()
-    capabilities_raw = form_data.get("capabilities", "").strip()
-    # Build full agent_id from org + name
-    agent_id = f"{org_id}::{agent_name}" if org_id and agent_name else ""
-    if not display_name:
-        display_name = agent_name.replace("-", " ").replace("_", " ").title()
-
-    cert_pem = form_data.get("cert_pem", "").strip()
-
-    form = {
-        "org_id": org_id,
-        "agent_name": agent_name,
-        "display_name": display_name,
-        "description": description,
-        "capabilities": capabilities_raw,
-        "cert_pem": cert_pem,
-    }
-
-    result = await db.execute(
-        select(OrganizationRecord)
-        .where(OrganizationRecord.status == "active")
-        .order_by(OrganizationRecord.org_id)
-    )
-    orgs = result.scalars().all()
-
-    # Org user can only register agents for their own org
-    if not session.is_admin and org_id != session.org_id:
-        return templates.TemplateResponse("agent_register.html",
-            _ctx(request, session, active="agents", form=form, orgs=orgs,
-                 error="You can only register agents for your own organization.", success=None))
-
-    if not org_id or not agent_name:
-        return templates.TemplateResponse("agent_register.html",
-            _ctx(request, session, active="agents", form=form, orgs=orgs,
-                 error="Organization and agent name are required.", success=None))
-
-    for val, lbl in [(org_id, "Organization ID"), (agent_name, "Agent name")]:
-        id_err = _validate_id(val, lbl)
-        if id_err:
-            return templates.TemplateResponse("agent_register.html",
-                _ctx(request, session, active="agents", form=form, orgs=orgs, error=id_err, success=None))
-
-    existing = await db.execute(
-        select(AgentRecord).where(AgentRecord.agent_id == agent_id)
-    )
-    if existing.scalar_one_or_none():
-        return templates.TemplateResponse("agent_register.html",
-            _ctx(request, session, active="agents", form=form, orgs=orgs,
-                 error=f"Agent '{agent_id}' already exists.", success=None))
-
-    # Audit F-B-2 — registering an agent on a sealed org implicitly
-    # auto-approves a binding and may pin a cert, both of which are
-    # tenant-identity-plane mutations. Require the re-auth gate.
-    target_org = await get_org_by_id(db, org_id)
-    if _is_sealed(target_org):
-        reauth_session = get_session(request)
-        if not reauth_session.has_reauth_scope(org_id):
-            return templates.TemplateResponse(
-                "agent_register.html",
-                _ctx(request, session, active="agents", form=form, orgs=orgs,
-                     error=(f"Org '{org_id}' is tenant-sealed. "
-                            f"Re-authenticate for this org before registering agents."),
-                     success=None),
-                status_code=403,
-            )
-
-    caps = [c.strip() for c in capabilities_raw.split(",") if c.strip()] if capabilities_raw else []
-    if len(caps) > _CAPABILITY_MAX_COUNT:
-        return templates.TemplateResponse("agent_register.html",
-            _ctx(request, session, active="agents", form=form, orgs=orgs,
-                 error=f"Maximum {_CAPABILITY_MAX_COUNT} capabilities allowed.", success=None))
-    for cap in caps:
-        if len(cap) > _CAPABILITY_MAX_LEN or not re.match(r"^[a-zA-Z0-9._:\-]+$", cap):
-            return templates.TemplateResponse("agent_register.html",
-                _ctx(request, session, active="agents", form=form, orgs=orgs,
-                     error=f"Invalid capability '{cap}'. Use alphanumeric, dots, colons, hyphens (max {_CAPABILITY_MAX_LEN} chars).",
-                     success=None))
-
-    await register_agent(
-        db, agent_id=agent_id, org_id=org_id,
-        display_name=display_name, capabilities=caps,
-        metadata={}, description=description,
-    )
-    await log_event(db, "registry.agent_registered", "ok",
-                    agent_id=agent_id, org_id=org_id,
-                    details=_sealed_mutation_details(
-                        target_org, extra={"capabilities": caps}))
-
-    # Auto-create and auto-approve binding. approved_by carries the
-    # sealed-vs-unsealed context for downstream forensics.
-    approved_by = (
-        "dashboard-admin-with-reauth" if _is_sealed(target_org) else "dashboard-admin"
-    )
-    existing_binding = await get_binding_by_org_agent(db, org_id, agent_id)
-    if existing_binding and existing_binding.status != "approved":
-        await approve_binding(db, existing_binding.id, approved_by=approved_by)
-    elif not existing_binding:
-        binding = await create_binding(db, org_id, agent_id, scope=caps)
-        await approve_binding(db, binding.id, approved_by=approved_by)
-
-    # If a certificate was provided, validate and pin it
-    cert_msg = ""
-    if cert_pem:
-        if "-----BEGIN CERTIFICATE-----" not in cert_pem:
-            cert_msg = " Certificate ignored: invalid PEM format."
-        else:
-            try:
-                from cryptography.x509 import load_pem_x509_certificate
-                from cryptography.x509.oid import NameOID
-                cert_obj = load_pem_x509_certificate(cert_pem.encode())
-                cn_attrs = cert_obj.subject.get_attributes_for_oid(NameOID.COMMON_NAME)
-                if not cn_attrs or cn_attrs[0].value != agent_id:
-                    cert_msg = f" Certificate ignored: CN '{cn_attrs[0].value if cn_attrs else '(none)'}' does not match agent ID."
-                else:
-                    # Verify against org CA if available
-                    org = await get_org_by_id(db, org_id)
-                    ca_ok = True
-                    if org and org.ca_certificate:
-                        try:
-                            from cryptography.hazmat.primitives.asymmetric import padding
-                            ca_cert = load_pem_x509_certificate(org.ca_certificate.encode())
-                            ca_cert.public_key().verify(
-                                cert_obj.signature, cert_obj.tbs_certificate_bytes,
-                                padding.PKCS1v15(), cert_obj.signature_hash_algorithm,
-                            )
-                        except Exception:
-                            ca_ok = False
-                            cert_msg = " Certificate ignored: not signed by organization CA."
-                    if ca_ok:
-                        await rotate_agent_cert(db, agent_id, cert_pem)
-                        await log_event(
-                            db, "registry.agent_cert_uploaded", "ok",
-                            agent_id=agent_id, org_id=org_id,
-                            details=_sealed_mutation_details(
-                                target_org, extra={"method": "register"}),
-                        )
-                        cert_msg = " Certificate pinned."
-            except Exception:
-                cert_msg = " Certificate ignored: could not parse PEM."
-
-    return templates.TemplateResponse("agent_register.html",
-        _ctx(request, session, active="agents", form={}, orgs=orgs, error=None,
-             success=f"Agent '{agent_id}' registered. Binding approved.{cert_msg}"))
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Manage Agent (unified settings page)
-# ─────────────────────────────────────────────────────────────────────────────
-
-@router.get("/agents/{agent_id:path}/manage", response_class=HTMLResponse)
-async def agent_manage_form(request: Request, agent_id: str, db: AsyncSession = Depends(get_db)):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    agent = (await db.execute(
-        select(AgentRecord).where(AgentRecord.agent_id == agent_id)
-    )).scalar_one_or_none()
-    if not agent:
-        return RedirectResponse(url="/dashboard/agents", status_code=303)
-    if not session.is_admin and agent.org_id != session.org_id:
-        return RedirectResponse(url="/dashboard/agents", status_code=303)
-    binding = await get_binding_by_org_agent(db, agent.org_id, agent_id)
-    ws_connected = ws_manager.is_connected(agent_id)
-    return templates.TemplateResponse("agent_manage.html",
-        _ctx(request, session, active="agents",
-             agent=agent, binding=binding, ws_connected=ws_connected,
-             error=None, success=None))
-
-
-@router.post("/agents/{agent_id:path}/manage", response_class=HTMLResponse)
-async def agent_manage_submit(request: Request, agent_id: str, db: AsyncSession = Depends(get_db)):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    if not await verify_csrf(request, session):
-        raise HTTPException(status_code=403, detail="Invalid CSRF token")
-    agent = (await db.execute(
-        select(AgentRecord).where(AgentRecord.agent_id == agent_id)
-    )).scalar_one_or_none()
-    if not agent:
-        return RedirectResponse(url="/dashboard/agents", status_code=303)
-    if not session.is_admin and agent.org_id != session.org_id:
-        return RedirectResponse(url="/dashboard/agents", status_code=303)
-
-    form_data = await request.form()
-    action = form_data.get("action", "")
-    binding = await get_binding_by_org_agent(db, agent.org_id, agent_id)
-    ws_connected = ws_manager.is_connected(agent_id)
-
-    def _render(error=None, success=None, status_code=200):
-        return templates.TemplateResponse(
-            "agent_manage.html",
-            _ctx(request, session, active="agents",
-                 agent=agent, binding=binding, ws_connected=ws_connected,
-                 error=error, success=success),
-            status_code=status_code,
-        )
-
-    # Audit F-B-2 — mutating an agent record on a sealed org requires the
-    # per-org re-auth scope. Check once here so both update_profile and
-    # upload_cert branches are covered.
-    owner_org = await get_org_by_id(db, agent.org_id)
-    if _is_sealed(owner_org) and action in ("update_profile", "upload_cert"):
-        session_live = get_session(request)
-        if not session_live.has_reauth_scope(agent.org_id):
-            return _render(
-                error=(f"Org '{agent.org_id}' is tenant-sealed. "
-                       f"Re-authenticate for this org before mutating its agents."),
-                status_code=403,
-            )
-
-    if action == "update_profile":
-        display_name = form_data.get("display_name", "").strip()
-        description = form_data.get("description", "").strip()
-        capabilities_raw = form_data.get("capabilities", "").strip()
-        caps = [c.strip() for c in capabilities_raw.split(",") if c.strip()] if capabilities_raw else []
-
-        if not display_name:
-            return _render(error="Display name is required.")
-
-        for cap in caps:
-            if len(cap) > 64 or not re.match(r"^[a-zA-Z0-9._:\-]+$", cap):
-                return _render(error=f"Invalid capability '{cap}'.")
-
-        agent.display_name = display_name
-        agent.description = description
-        agent.capabilities_json = _json.dumps(caps)
-
-        # Update binding scope to match capabilities
-        if binding:
-            binding.scope_json = _json.dumps(caps)
-
-        await db.commit()
-        await log_event(
-            db, "registry.agent_updated", "ok",
-            agent_id=agent_id, org_id=agent.org_id,
-            details=_sealed_mutation_details(
-                owner_org,
-                extra={"fields": ["display_name", "description", "capabilities"]},
-            ),
-        )
-        return _render(success="Agent profile updated.")
-
-    elif action == "upload_cert":
-        cert_pem = form_data.get("cert_pem", "").strip()
-        if not cert_pem or "-----BEGIN CERTIFICATE-----" not in cert_pem:
-            return _render(error="Invalid certificate. Paste a valid PEM certificate.")
-        try:
-            from cryptography.x509 import load_pem_x509_certificate
-            from cryptography.x509.oid import NameOID
-            cert = load_pem_x509_certificate(cert_pem.encode())
-        except Exception:
-            return _render(error="Could not parse the certificate.")
-        cn_attrs = cert.subject.get_attributes_for_oid(NameOID.COMMON_NAME)
-        if not cn_attrs or cn_attrs[0].value != agent_id:
-            return _render(error=f"Certificate CN does not match agent ID '{agent_id}'.")
-        # Verify against org CA if available
-        org = await get_org_by_id(db, agent.org_id)
-        if org and org.ca_certificate:
-            try:
-                from cryptography.hazmat.primitives.asymmetric import padding as _pad, ec as _ec
-                ca_cert = load_pem_x509_certificate(org.ca_certificate.encode())
-                ca_pub = ca_cert.public_key()
-                from cryptography.hazmat.primitives.asymmetric import rsa as _rsa
-                if isinstance(ca_pub, _rsa.RSAPublicKey):
-                    ca_pub.verify(cert.signature, cert.tbs_certificate_bytes,
-                                  _pad.PKCS1v15(), cert.signature_hash_algorithm)
-                elif isinstance(ca_pub, _ec.EllipticCurvePublicKey):
-                    ca_pub.verify(cert.signature, cert.tbs_certificate_bytes,
-                                  _ec.ECDSA(cert.signature_hash_algorithm))
-            except Exception:
-                return _render(error="Certificate not signed by organization CA.")
-        new_thumbprint = await rotate_agent_cert(db, agent_id, cert_pem)
-        await log_event(db, "registry.agent_cert_uploaded", "ok",
-                        agent_id=agent_id, org_id=agent.org_id,
-                        details=_sealed_mutation_details(owner_org))
-        # Refresh agent
-        agent = (await db.execute(
-            select(AgentRecord).where(AgentRecord.agent_id == agent_id)
-        )).scalar_one_or_none()
-        return _render(success=f"Certificate uploaded. Thumbprint: {new_thumbprint[:16]}...")
-
-    return _render(error="Unknown action.")
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Delete Agent
-# ─────────────────────────────────────────────────────────────────────────────
-
-@router.post("/agents/{agent_id:path}/delete", response_class=HTMLResponse)
-async def agent_delete(request: Request, agent_id: str, db: AsyncSession = Depends(get_db)):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    if not session.is_admin:
-        return RedirectResponse(url="/dashboard/agents", status_code=303)
-    if not await verify_csrf(request, session):
-        return RedirectResponse(url="/dashboard/agents", status_code=303)
-
-    agent = await db.execute(select(AgentRecord).where(AgentRecord.agent_id == agent_id))
-    record = agent.scalar_one_or_none()
-    if record:
-        # Audit F-B-2 — deleting an agent on a sealed org revokes its
-        # binding and nukes its directory entry. Gate on re-auth.
-        owner_org = await get_org_by_id(db, record.org_id)
-        await _require_sealed_reauth(request, owner_org)
-
-        # Revoke binding if exists
-        binding = await get_binding_by_org_agent(db, record.org_id, agent_id)
-        if binding and binding.status != "revoked":
-            await revoke_binding(db, binding.id)
-
-        # Delete the agent
-        await db.delete(record)
-        await db.commit()
-        await log_event(db, "registry.agent_deleted", "ok",
-                        agent_id=agent_id, org_id=record.org_id,
-                        details=_sealed_mutation_details(owner_org))
-
-    return RedirectResponse(url="/dashboard/agents", status_code=303)
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Agent Detail — Developer Portal Page
-# ─────────────────────────────────────────────────────────────────────────────
-
-# ``_broker_url_from_request`` lives in ``app.dashboard._helpers`` since
-# F-B-202 PR-1.
-
-
-@router.get("/agents/{agent_id:path}", response_class=HTMLResponse)
-async def agent_detail(request: Request, agent_id: str,
-                       db: AsyncSession = Depends(get_db)):
-    """Developer portal page for a single agent."""
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-
-    agent = (await db.execute(
-        select(AgentRecord).where(AgentRecord.agent_id == agent_id)
-    )).scalar_one_or_none()
-    if not agent:
-        raise HTTPException(status_code=404, detail="Agent not found")
-
-    if not session.is_admin and agent.org_id != session.org_id:
-        return RedirectResponse(url="/dashboard/agents", status_code=303)
-
-    # Binding
-    binding = await get_binding_by_org_agent(db, agent.org_id, agent_id)
-
-    # WebSocket status
-    ws_connected = ws_manager.is_connected(agent_id)
-
-    # Certificate expiry
-    cert_expiry = None
-    if agent.cert_pem:
-        try:
-            from cryptography.x509 import load_pem_x509_certificate
-            cert = load_pem_x509_certificate(agent.cert_pem.encode())
-            cert_expiry = cert.not_valid_after_utc.strftime("%Y-%m-%d")
-        except Exception:
-            pass
-
-    # Recent audit events
-    from app.db.audit import AuditLog
-    q = (select(AuditLog)
-         .where(AuditLog.agent_id == agent_id)
-         .order_by(AuditLog.id.desc())
-         .limit(10))
-    audit_events = (await db.execute(q)).scalars().all()
-
-    broker_url = _broker_url_from_request(request)
-
-    return templates.TemplateResponse("agent_detail.html",
-        _ctx(request, session, active="agents",
-             agent=agent, binding=binding, ws_connected=ws_connected,
-             cert_expiry=cert_expiry, audit_events=audit_events,
-             broker_url=broker_url))
-
-
-@router.get("/agents/{agent_id:path}/upload-cert", response_class=HTMLResponse)
-async def agent_upload_cert_form(request: Request, agent_id: str,
-                                  db: AsyncSession = Depends(get_db)):
-    """Show the standalone certificate upload form."""
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    agent = (await db.execute(
-        select(AgentRecord).where(AgentRecord.agent_id == agent_id)
-    )).scalar_one_or_none()
-    if not agent:
-        return RedirectResponse(url="/dashboard/agents", status_code=303)
-    if not session.is_admin and agent.org_id != session.org_id:
-        return RedirectResponse(url="/dashboard/agents", status_code=303)
-    return templates.TemplateResponse("cert_upload.html",
-        _ctx(request, session, active="agents", agent=agent, error=None, success=None))
-
-
-@router.post("/agents/{agent_id:path}/upload-cert", response_class=HTMLResponse)
-async def agent_upload_cert(request: Request, agent_id: str,
-                            db: AsyncSession = Depends(get_db)):
-    """Upload an externally signed agent certificate (BYOCA production flow)."""
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    if not await verify_csrf(request, session):
-        raise HTTPException(status_code=403, detail="Invalid CSRF token")
-
-    agent = (await db.execute(
-        select(AgentRecord).where(AgentRecord.agent_id == agent_id)
-    )).scalar_one_or_none()
-    if not agent:
-        raise HTTPException(status_code=404, detail="Agent not found")
-    if not session.is_admin and agent.org_id != session.org_id:
-        raise HTTPException(status_code=403, detail="Access denied")
-
-    # Audit F-B-2 — pinning a cert on a sealed org's agent rewrites an
-    # identity. Gate on the per-org re-auth scope.
-    owner_org = await get_org_by_id(db, agent.org_id)
-    await _require_sealed_reauth(request, owner_org)
-
-    form_data = await request.form()
-    cert_pem = form_data.get("cert_pem", "").strip()
-
-    # Helper to re-render the upload page with an error
-    def _render_error(error_msg):
-        return templates.TemplateResponse("cert_upload.html",
-            _ctx(request, session, active="agents",
-                 agent=agent, error=error_msg, success=None))
-
-    if not cert_pem or "-----BEGIN CERTIFICATE-----" not in cert_pem:
-        return _render_error("Invalid certificate. Paste a valid PEM certificate.")
-
-    # Parse and validate the certificate
-    try:
-        from cryptography.x509 import load_pem_x509_certificate
-        cert = load_pem_x509_certificate(cert_pem.encode())
-    except Exception:
-        return _render_error("Could not parse the certificate. Ensure it is valid PEM format.")
-
-    # Verify CN matches agent_id
-    from cryptography.x509.oid import NameOID
-    cn_attrs = cert.subject.get_attributes_for_oid(NameOID.COMMON_NAME)
-    if not cn_attrs or cn_attrs[0].value != agent_id:
-        return _render_error(
-            f"Certificate CN '{cn_attrs[0].value if cn_attrs else '(none)'}' "
-            f"does not match agent ID '{agent_id}'.")
-
-    # Verify cert is signed by the org's CA (if CA is uploaded). Use the
-    # org record we already fetched above for the sealed check.
-    if owner_org and owner_org.ca_certificate:
-        try:
-            from cryptography.x509 import load_pem_x509_certificate as _load_cert
-            from cryptography.hazmat.primitives.asymmetric import padding
-            ca_cert = _load_cert(owner_org.ca_certificate.encode())
-            ca_cert.public_key().verify(
-                cert.signature,
-                cert.tbs_certificate_bytes,
-                padding.PKCS1v15(),
-                cert.signature_hash_algorithm,
-            )
-        except Exception:
-            return _render_error(
-                "Certificate signature verification failed. "
-                "The certificate must be signed by your organization's CA.")
-
-    # Pin the certificate
-    from app.registry.store import rotate_agent_cert
-    new_thumbprint = await rotate_agent_cert(db, agent_id, cert_pem)
-
-    await log_event(
-        db, "registry.agent_cert_uploaded", "ok",
-        agent_id=agent_id, org_id=agent.org_id,
-        details=_sealed_mutation_details(owner_org, extra={"method": "upload"}),
-    )
-
-    # Re-read the agent to show updated thumbprint
-    agent = (await db.execute(
-        select(AgentRecord).where(AgentRecord.agent_id == agent_id)
-    )).scalar_one_or_none()
-    return templates.TemplateResponse("cert_upload.html",
-        _ctx(request, session, active="agents", agent=agent, error=None,
-             success=f"Certificate uploaded. Thumbprint: {new_thumbprint[:16]}..."))
-
-
-@router.post("/agents/{agent_id:path}/credentials")
-async def agent_credentials_download(request: Request, agent_id: str,
-                                     db: AsyncSession = Depends(get_db)):
-    """Generate and download credentials-only bundle (cert + key + env)."""
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    if not await verify_csrf(request, session):
-        raise HTTPException(status_code=403, detail="Invalid CSRF token")
-
-    agent = (await db.execute(
-        select(AgentRecord).where(AgentRecord.agent_id == agent_id)
-    )).scalar_one_or_none()
-    if not agent:
-        raise HTTPException(status_code=404, detail="Agent not found")
-    if not session.is_admin and agent.org_id != session.org_id:
-        raise HTTPException(status_code=403, detail="Access denied")
-
-    org_id = agent.org_id
-
-    # Audit F-B-2 — this endpoint mints fresh (key, cert) material for an
-    # agent and pins the cert thumbprint in the DB. If the owning org is
-    # sealed, the admin must clear the per-org re-auth gate first.
-    owner_org = await get_org_by_id(db, org_id)
-    await _require_sealed_reauth(request, owner_org)
-
-    # Load org CA
-    from cryptography.hazmat.primitives.serialization import load_pem_private_key
-    from cryptography.x509 import load_pem_x509_certificate
-
-    certs_dir = pathlib.Path(__file__).parent.parent.parent / "certs"
-    org_ca_key_path = certs_dir / org_id / "ca-key.pem"
-    org_ca_cert_path = certs_dir / org_id / "ca.pem"
-
-    if not org_ca_key_path.exists() or not org_ca_cert_path.exists():
-        raise HTTPException(status_code=500,
-            detail=f"Org CA not found for '{org_id}'. Upload CA certificate first.")
-
-    org_ca_key = load_pem_private_key(org_ca_key_path.read_bytes(), password=None)
-    org_ca_cert = load_pem_x509_certificate(org_ca_cert_path.read_bytes())
-
-    # Generate cert + key
-    key_pem, cert_pem = _generate_agent_cert(agent_id, org_id, org_ca_key, org_ca_cert)
-
-    # Pin cert in DB
-    from app.registry.store import rotate_agent_cert
-    await rotate_agent_cert(db, agent_id, cert_pem.decode())
-
-    broker_url = _broker_url_from_request(request)
-
-    # Minimal env — just connection essentials
-    env_content = (
-        f"# Cullis — credentials\n"
-        f"# Generated: {datetime.datetime.now(datetime.timezone.utc).isoformat()}\n"
-        f"BROKER_URL={broker_url}\n"
-        f"AGENT_ID={agent_id}\n"
-        f"ORG_ID={org_id}\n"
-        f"CAPABILITIES={','.join(agent.capabilities)}\n"
-    )
-
-    # Build credentials-only zip
-    buf = io.BytesIO()
-    with zipfile.ZipFile(buf, "w", zipfile.ZIP_DEFLATED) as zf:
-        zf.writestr("agent.pem", cert_pem)
-        zf.writestr("agent-key.pem", key_pem)
-        zf.writestr("agent.env", env_content)
-
-    buf.seek(0)
-    safe_name = agent_id.replace("::", "__")
-
-    await log_event(db, "registry.agent_credentials_generated", "ok",
-                    agent_id=agent_id, org_id=org_id,
-                    details=_sealed_mutation_details(owner_org))
-
-    return StreamingResponse(
-        buf,
-        media_type="application/zip",
-        headers={"Content-Disposition": f'attachment; filename="{safe_name}-credentials.zip"'},
-    )
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Download Agent Bundle (zip with cert, key, env, scripts, start.sh)
-# ─────────────────────────────────────────────────────────────────────────────
-
-# ``_generate_agent_cert`` lives in ``app.dashboard._helpers`` since
-# F-B-202 PR-1.
-
-
-@router.post("/agents/{agent_id:path}/bundle")
-async def agent_bundle_download(request: Request, agent_id: str, db: AsyncSession = Depends(get_db)):
-    """Generate and download a deploy bundle (zip) for an agent."""
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    if not await verify_csrf(request, session):
-        raise HTTPException(status_code=403, detail="Invalid CSRF token")
-
-    # Load agent record
-    agent = (await db.execute(
-        select(AgentRecord).where(AgentRecord.agent_id == agent_id)
-    )).scalar_one_or_none()
-    if not agent:
-        raise HTTPException(status_code=404, detail="Agent not found")
-
-    # Org user can only download own agents
-    if not session.is_admin and agent.org_id != session.org_id:
-        raise HTTPException(status_code=403, detail="Access denied")
-
-    org_id = agent.org_id
-    caps = agent.capabilities
-
-    # Audit F-B-2 — this endpoint mints fresh (key, cert) material and
-    # pins the thumbprint in the DB. Same threat model as /credentials
-    # above — sealed orgs need the per-org re-auth gate.
-    owner_org = await get_org_by_id(db, org_id)
-    await _require_sealed_reauth(request, owner_org)
-
-    # Load org CA to sign agent cert
-    from cryptography.hazmat.primitives.serialization import load_pem_private_key
-    from cryptography.x509 import load_pem_x509_certificate
-
-    certs_dir = pathlib.Path(__file__).parent.parent.parent / "certs"
-    org_ca_key_path = certs_dir / org_id / "ca-key.pem"
-    org_ca_cert_path = certs_dir / org_id / "ca.pem"
-
-    if not org_ca_key_path.exists() or not org_ca_cert_path.exists():
-        raise HTTPException(status_code=500,
-            detail=f"Org CA not found for '{org_id}'. Run join.py first.")
-
-    org_ca_key = load_pem_private_key(org_ca_key_path.read_bytes(), password=None)
-    org_ca_cert = load_pem_x509_certificate(org_ca_cert_path.read_bytes())
-
-    # Generate agent cert + key
-    key_pem, cert_pem = _generate_agent_cert(agent_id, org_id, org_ca_key, org_ca_cert)
-
-    # Pin cert in DB
-    from app.registry.store import rotate_agent_cert
-    await rotate_agent_cert(db, agent_id, cert_pem.decode())
-
-    # Determine broker URL from request
-    scheme = request.headers.get("x-forwarded-proto", request.url.scheme)
-    host = request.headers.get("x-forwarded-host", request.url.hostname)
-    port = request.url.port
-    if scheme == "https" and port and port != 443:
-        broker_url = f"{scheme}://{host}:{port}"
-    elif scheme == "http" and port and port != 80:
-        broker_url = f"{scheme}://{host}:{port}"
-    else:
-        broker_url = f"{scheme}://{host}"
-
-    # Build .env content
-    safe_name = agent_id.replace("::", "__")
-    env_content = (
-        f"# Cullis — deploy bundle\n"
-        f"# Generated: {datetime.datetime.now(datetime.timezone.utc).isoformat()}\n"
-        f"BROKER_URL={broker_url}\n"
-        f"AGENT_ID={agent_id}\n"
-        f"ORG_ID={org_id}\n"
-        f"DISPLAY_NAME={agent.display_name}\n"
-        f"AGENT_CERT_PATH=./{safe_name}.pem\n"
-        f"AGENT_KEY_PATH=./{safe_name}-key.pem\n"
-        f"ORG_SECRET={org_id}\n"
-        f"CAPABILITIES={','.join(caps)}\n"
-        f"POLL_INTERVAL=2\n"
-        f"MAX_TURNS=20\n"
-        f"\n"
-        f"# LLM backend\n"
-        f"LLM_MODEL=claude-sonnet-4-6\n"
-        f"ANTHROPIC_API_KEY=\n"
-    )
-
-    # Build start.sh — authenticates agent and connects to the network
-    start_sh = (
-        "#!/usr/bin/env bash\n"
-        "set -euo pipefail\n"
-        'cd "$(dirname "$0")"\n'
-        'echo "Connecting agent to ATN broker..."\n'
-        "python agent_node.py --config agent.env \"$@\"\n"
-    )
-
-    # Read the demo scripts to include in bundle
-    demo_dir = pathlib.Path(__file__).parent.parent.parent / "demo"
-    sdk_path = pathlib.Path(__file__).parent.parent.parent / "agents" / "sdk.py"
-
-    # Build the zip
-    buf = io.BytesIO()
-    with zipfile.ZipFile(buf, "w", zipfile.ZIP_DEFLATED) as zf:
-        # Cert and key
-        zf.writestr(f"{safe_name}.pem", cert_pem)
-        zf.writestr(f"{safe_name}-key.pem", key_pem)
-
-        # Env
-        zf.writestr("agent.env", env_content)
-
-        # start.sh (executable)
-        info = zipfile.ZipInfo("start.sh")
-        info.external_attr = 0o755 << 16
-        zf.writestr(info, start_sh)
-
-        # SDK
-        if sdk_path.exists():
-            zf.writestr("agents/sdk.py", sdk_path.read_text())
-            zf.writestr("agents/__init__.py", "")
-
-        # All demo scripts — the user decides what to run
-        for name in ("agent_node.py", "buyer_agent.py", "supplier_agent.py",
-                      "inventory_watcher.py", "inventory.json"):
-            path = demo_dir / name
-            if path.exists():
-                zf.writestr(name, path.read_text())
-
-    buf.seek(0)
-    filename = f"{safe_name}-bundle.zip"
-
-    await log_event(db, "registry.agent_bundle_downloaded", "ok",
-                    agent_id=agent_id, org_id=org_id,
-                    details={"source": "dashboard"})
-
-    return StreamingResponse(
-        buf,
-        media_type="application/zip",
-        headers={"Content-Disposition": f'attachment; filename="{filename}"'},
-    )
-
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Rotate Certificate
-# ─────────────────────────────────────────────────────────────────────────────
-
-@router.get("/agents/{agent_id:path}/rotate-cert", response_class=HTMLResponse)
-async def cert_rotate_form(request: Request, agent_id: str, db: AsyncSession = Depends(get_db)):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-
-    agent = (await db.execute(select(AgentRecord).where(AgentRecord.agent_id == agent_id))).scalar_one_or_none()
-    if not agent:
-        return RedirectResponse(url="/dashboard/agents", status_code=303)
-    if not session.is_admin and agent.org_id != session.org_id:
-        return RedirectResponse(url="/dashboard/agents", status_code=303)
-
-    return templates.TemplateResponse("cert_rotate.html",
-        _ctx(request, session, active="agents", agent=agent, error=None, success=None))
-
-
-@router.post("/agents/{agent_id:path}/rotate-cert", response_class=HTMLResponse)
-async def cert_rotate_submit(request: Request, agent_id: str, db: AsyncSession = Depends(get_db)):
-    session = require_login(request)
-    if isinstance(session, RedirectResponse):
-        return session
-    if not await verify_csrf(request, session):
-        return RedirectResponse(url="/dashboard/agents", status_code=303)
-
-    agent = (await db.execute(select(AgentRecord).where(AgentRecord.agent_id == agent_id))).scalar_one_or_none()
-    if not agent:
-        return RedirectResponse(url="/dashboard/agents", status_code=303)
-    if not session.is_admin and agent.org_id != session.org_id:
-        return RedirectResponse(url="/dashboard/agents", status_code=303)
-
-    # Audit F-B-2 — rotating a cert on a sealed org's agent rewrites its
-    # identity. Gate on the per-org re-auth scope.
-    owner_org = await get_org_by_id(db, agent.org_id)
-    await _require_sealed_reauth(request, owner_org)
-
-    form_data = await request.form()
-    cert_pem = form_data.get("certificate", "").strip()
-
-    if not cert_pem:
-        return templates.TemplateResponse("cert_rotate.html",
-            _ctx(request, session, active="agents", agent=agent,
-                 error="Certificate PEM is required.", success=None))
-
-    # Validate the certificate
-    from cryptography import x509 as crypto_x509
-    from cryptography.exceptions import InvalidSignature
-    from cryptography.hazmat.primitives.asymmetric import padding
-    from cryptography.x509.oid import NameOID
-
-    try:
-        cert = crypto_x509.load_pem_x509_certificate(cert_pem.encode())
-    except Exception:
-        return templates.TemplateResponse("cert_rotate.html",
-            _ctx(request, session, active="agents", agent=agent,
-                 error="Invalid PEM certificate.", success=None))
-
-    # Verify CN matches agent_id
-    cn_attrs = cert.subject.get_attributes_for_oid(NameOID.COMMON_NAME)
-    if not cn_attrs or cn_attrs[0].value != agent_id:
-        return templates.TemplateResponse("cert_rotate.html",
-            _ctx(request, session, active="agents", agent=agent,
-                 error=f"Certificate CN does not match agent '{agent_id}'.",
-                 success=None))
-
-    # Verify signed by org CA (if CA is configured). Reuse owner_org from
-    # the sealed-check step above so we don't hit the DB twice.
-    if owner_org and owner_org.ca_certificate:
-        try:
-            org_ca = crypto_x509.load_pem_x509_certificate(owner_org.ca_certificate.encode())
-            org_ca.public_key().verify(
-                cert.signature, cert.tbs_certificate_bytes,
-                padding.PKCS1v15(), cert.signature_hash_algorithm,
-            )
-        except InvalidSignature:
-            return templates.TemplateResponse("cert_rotate.html",
-                _ctx(request, session, active="agents", agent=agent,
-                     error="Certificate is not signed by the organization CA.", success=None))
-        except Exception:
-            return templates.TemplateResponse("cert_rotate.html",
-                _ctx(request, session, active="agents", agent=agent,
-                     error="Certificate verification failed. Please check the certificate is valid and signed by the organization CA.", success=None))
-
-    old_thumbprint = agent.cert_thumbprint
-    new_thumbprint = await rotate_agent_cert(db, agent_id, cert_pem)
-
-    await log_event(db, "agent.cert_rotated", "ok",
-                    agent_id=agent_id, org_id=agent.org_id,
-                    details=_sealed_mutation_details(
-                        owner_org,
-                        extra={"old_thumbprint": old_thumbprint,
-                               "new_thumbprint": new_thumbprint},
-                    ))
-
-    return templates.TemplateResponse("cert_rotate.html",
-        _ctx(request, session, active="agents", agent=agent, error=None,
-             success=f"Certificate rotated. New thumbprint: {new_thumbprint[:16]}…"))
 
 
 # ─────────────────────────────────────────────────────────────────────────────
